Redesigned scrape.py
scrape.py works more efficiently with mapping.py and user inputs. User can now go from an external file and pull the dataframe directly from the pgatour.com stats pages. Program currently works better if stat_id is declared as it is difficult to match text directly. Future plan is to update the case matching for the statistic of interest.