Modified format of meta data storage
Created a meta_data.py file in the pga_data directory to house the meta data extracted from the pgatour.com website. This allows for a more complete and succinct data retrieval and analysis process.

This commit includes the tourney_meta variable within the meta_data.py file and a temporary stat_meta variable. The goal for next commit is to fill the stat_meta with all stats utilizing the functions in mapping.py.