Organized mapping.py into classes
For efficiency, mapping.py was broken into classes depending on the tasks that you are performing. The mapping class is for retreiving metadata for each statistic and the make_meta class is for generating the json files used to generate the url for each statistic.  make_meta is mostly for storage of functions that were generated for a specific purpose previously but might be recalled later.

Next step is to test scrape with the mapping.mapping class and determine necessary improvements going forward.